requirement.txt
attrs==21.2.0
autopep8==1.5.7
aws-encryption-sdk==2.3.0
boto3==1.17.102
botocore==1.20.102
cffi==1.14.5
cryptography==3.4.7
jmespath==0.10.0
pycodestyle==2.7.0
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.4
s3transfer==0.4.2
six==1.16.0
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
